New Building Pallette System
Implemented a building palette similar to the shop in Hay Day.

- press "All Buildings" to open up shop
- drag the item from it's item slot
- dragging outside of the shop will spawn the item in-world which u can place.
- close button will turn off the palette.

Well about this week, It's not gonna be much. I've been lazing for most of this month, crawling towards the end. This may not change, but I will somehow finish the prototype by this month.

I don't think there will be a lot of work left to do.